requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pillow==9.5.0
pydicom==2.4.1
tifffile==2023.9.26
numpy==1.25.0
opencv-python==4.8.0.76
reportlab==4.0.4
pdf2image==1.16.3
typing-extensions==4.8.0
requests==2.31.0
python-multipart==0.0.6
slowapi
pyjwt[crypto]
